Created client prototype with working PUSH,PULL,LIST operations and socket interfacing. Created mock_sock for client testing. Modified message send/receive for socket interconnection.

+++ include/utils.h
@@ -9,18 +9,18 @@ int get_timestamp(char *buff,size_t size,time_t tm);
 
 /* Reads a ```req_size``` chunk from ```fd``` into buffer
     Returns size read ,-1 on error.*/
-ssize_t read_buff(char* buffer,ssize_t req_size,int fd);
+ssize_t read_buff(char* buffer,size_t req_size,int fd);
 
 /* Write a ```req_size``` chunk from buffer into fd/
     Returns size written , -1 on error.*/
 
-ssize_t write_buff(char* buff,ssize_t req_size,int fd);
+ssize_t write_buff(char* buff,size_t req_size,int fd);
 /* Receive a messsage from ```pipe_in```in ```buff```.
     Returns size received, -1 on error.*/
 ssize_t receive_msg(int pipe_in,char* buff);
 
 /* Send a messsage stored in ```buff``` through ```pipe_out```.
     Returns -1 on error.*/
-int send_msg(char *msg,ssize_t size,int pipe_out);
+ssize_t send_msg(char *msg,size_t size,int pipe_out);
 
 #endif
